feat: add @vercel/og image generation and full OpenGraph metadata
- Remove output:'export' from next.config.ts — static export is
  incompatible with Edge Functions (required by @vercel/og)
- Restore framework:"nextjs" in vercel.json; drop outputDirectory:"dist"
- Add @vercel/og ^0.6.3 to dependencies
- Add /og Edge route (src/app/og/route.tsx): generates 1200×630 PNG
  from ?title, ?description, ?label, ?tags params. Design: dark navy
  (#070D1E) with dot-grid, blue-cyan radial glows, purple accent,
  ghost icon watermark, squircle logo, gradient separator, tag badges
- Add src/lib/og.ts helper to build /og URLs with URLSearchParams
- Split every page into server wrapper (page.tsx, exports Metadata) +
  client component (_page.tsx / _home.tsx, keeps 'use client')
- layout.tsx: metadataBase, title template, robots, openGraph with
  alternateLocale for ru_RU / pl_PL, Twitter card
- Per-page metadata: unique title, description, label, tags for
  home, docs, download, help, legal, support pages
